+++ components/display/avatar/Avatar.mjs
@@ -63,12 +63,12 @@ export default class Avatar extends AbstractComponent {
     }
 
     static _rawStylingRules = [
-        `.avatar{--avatar-size:2em;--avatar-bg:var(--primary-90);--avatar-fg:var(--primary-10);--avatar-border-radius:var(--border-roundness);display:flex;align-items:center;justify-content:center;line-height:1;overflow:hidden;width:var(--avatar-size);height:var(--avatar-size);background-color:var(--avatar-bg);color:var(--avatar-fg);font-size:1.25em;border-radius:var(--avatar-border-radius);user-select:none;margin:var(--padding-sm);background-size:contain;background-position:center;background-repeat:no-repeat}`,
+        `.avatar{--avatar-size:2em;--avatar-bg:var(--primary-90);--avatar-fg:var(--primary-10);--avatar-border-radius:var(--border-roundness);display:flex;align-items:center;justify-content:center;line-height:1;overflow:hidden;width:var(--avatar-size);height:var(--avatar-size);background-color:var(--avatar-bg);color:var(--avatar-fg);font-size:1.25em;border-radius:var(--avatar-border-radius);user-select:none;margin:var(--padding-sm);background-size:contain;background-position:center;background-repeat:no-repeat;box-shadow:var(--box-shadow)}`,
         `.avatar.circular{--avatar-border-radius:50%}`,
         `.avatar.square{--avatar-border-radius:0}`,
         `.avatar.rounded{--avatar-border-radius:var(--border-roundness)}`,
         `.avatar.kw-{{theme}}{--avatar-bg:var(--{{theme}}-90);--avatar-fg:var(--{{theme}}-10)}`,
-        `.avatar.inline{margin:0;display:inline-flex;vertical-align:middle;margin-inline:var(--letter-spacing) var(--letter-spacing)}`,
+        `.avatar.inline{margin:0;display:inline-flex;vertical-align:baseline;margin-inline:var(--letter-spacing) var(--letter-spacing)}`,
         `.avatar-huge{font-size:3.739em}`,
         `.avatar-xxxl{font-size:3.324em}`,
         `.avatar-xxl{font-size:2.978em}`,